Sistema de indicadores proyecto 1/alcaldia_backend/requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0

# Autenticación JWT
djangorestframework-simplejwt==5.3.0

# Base de datos PostgreSQL
psycopg2-binary==2.9.9

# CORS para frontend separado
django-cors-headers==4.3.1

# Manejo de variables de entorno
python-decouple==3.8

# Procesamiento de archivos Excel
openpyxl==3.1.2
pandas==2.1.3

# Reportes PDF
reportlab==4.0.7
matplotlib==3.8.2

# Validación de archivos
python-magic==0.4.27

# Utilidades
pillow==10.1.0
requests==2.31.0

# Desarrollo y testing
pytest==7.4.3
pytest-django==4.7.0
coverage==7.3.2

# Producción
gunicorn==21.2.0
whitenoise==6.6.0